llm/requirements.txt
python-dotenv==1.0.0
transformers==4.34.1
accelerate==0.24.0
torch==2.1.0
fastapi==0.104.0
openai==0.28.1
